backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
sentence-transformers
openai
redis
rq
fakeredis
python-dotenv
